Replicate serialization issue
While deserializing a Java object, it can occur to have other calls to fetch other objects in order to assemble the original object composition.
This appears to be a resource starvation issue (number of client threds = CPU cores = 8 in my case).

If this is the case, then deserialization should not depend on the number of CPU cores or available threads.