Keeping current directory structure
Although simpler in terms of number of required changes, the fact is that maven only accepts a test directory per project. :cry: The main problem with this approach is that the developer will have to switch between modes at least once in order for the IDE to detect the possible test source directories. :sob: